backend/pyproject.toml
[project]
name = "quantum-katas"
version = "0.1.0"
description = "Interactive quantum computing learning platform with kata-style exercises"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "cirq-core>=1.4",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
    "httpx>=0.28",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

# ==============================================================================
# Ruff
# ==============================================================================
[tool.ruff]
target-version = "py312"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    "F",      # Pyflakes
    "E", "W", # pycodestyle
    "C90",    # mccabe (複雑度)
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "YTT",    # flake8-2020
    "ANN",    # flake8-annotations
    "ASYNC",  # flake8-async
    "S",      # flake8-bandit (セキュリティ)
    "BLE",    # flake8-blind-except
    "B",      # flake8-bugbear
    "A",      # flake8-builtins
    "COM",    # flake8-commas
    "C4",     # flake8-comprehensions
    "DTZ",    # flake8-datetimez
    "T10",    # flake8-debugger
    "EM",     # flake8-errmsg
    "FA",     # flake8-future-annotations
    "ISC",    # flake8-implicit-str-concat
    "ICN",    # flake8-import-conventions
    "LOG",    # flake8-logging
    "G",      # flake8-logging-format
    "PIE",    # flake8-pie
    "T20",    # flake8-print
    "PT",     # flake8-pytest-style
    "Q",      # flake8-quotes
    "RSE",    # flake8-raise
    "RET",    # flake8-return
    "SLF",    # flake8-self
    "SLOT",   # flake8-slots
    "SIM",    # flake8-simplify
    "TID",    # flake8-tidy-imports
    "TCH",    # flake8-type-checking
    "ARG",    # flake8-unused-arguments
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (コメントアウトコード検出)
    "PL",     # pylint
    "TRY",    # tryceratops
    "FLY",    # flynt
    "PERF",   # perflint
    "FURB",   # refurb
    "RUF",    # Ruff 固有ルール
]
ignore = [
    "ANN401",  # Any の使用（必要な場合あり、レビューで判断）
    "COM812",  # trailing comma（formatter と競合）
    "ISC001",  # implicit string concat（formatter と競合）
    "S101",    # assert の使用（テストで必要）
    "PLR0913", # 引数が多すぎる（DI で避けられない場合あり）
    "TRY003",  # 長いエラーメッセージ（実用性優先）
]

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.pylint]
max-args = 8
max-returns = 6
max-branches = 12

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101",    # assert 許可
    "ANN",     # テストの型注釈は任意
    "ARG",     # fixture の未使用引数
    "PLR2004", # マジックナンバー許可
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

# ==============================================================================
# pytest
# ==============================================================================
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-fail-under=80",
]
markers = [
    "slow: 実行に時間がかかるテスト",
    "integration: 外部依存のある統合テスト",
]
asyncio_mode = "auto"

# ==============================================================================
# Coverage
# ==============================================================================
[tool.coverage.run]
branch = true
source = ["src"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
    "@overload",
    "raise NotImplementedError",
]
